Make code quality improvements * Add more file comments * Replace tm printing method with std::put_time() * fix #53 * fix #52

+++ lathraea-rhodopaea/src/front-end/userManagementMenu.cpp
@@ -1,3 +1,7 @@
+/*! @file userManagementMenu.cpp
+*   @brief A source file for the user management menu handaling system.
+*/
+
 #include <front-end/userManagementMenu.h>
 #include <front-end/output.h>
 #include <front-end/graphics.h>
@@ -249,7 +253,7 @@ void viewAllUsers(AccountManager* accountManager)
 		case (int)ARROW_KEYS::KEY_ENTER:
 			clearConsole();
 			printFullyOpenedBook();
-			showUserInfo(allAccounts[int(selectedOption)- 1], accountManager);
+			showUserInfo(allAccounts[__int64(selectedOption)- 1], accountManager);
 		case (int)ARROW_KEYS::KEY_ESC:
 			// Return to main menu
 			clearConsole();
